requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
google-generativeai==0.3.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
Werkzeug==3.0.1
alembic==1.13.1
gunicorn==21.2.0
pytest==7.4.3
